refactor(tools): eliminate duplication, split common.ts, unify shared utilities
P0 - Eliminate code duplication:
- Add tools/shared/net.ts with canonical reserveOpenPort
- Add fetchJson/fetchJsonResponse to tools/shared/http.ts
- Add tools/tauri/smoke-harness.ts with shared CRUD smoke test logic
- Rewrite 3 smoke tests to use smoke harness (-70% lines)

P1 - Improve module structure:
- Extract tools/tauri/constants.ts (69 lines) from common.ts
- Extract tools/tauri/node-runtime.ts (166 lines) from common.ts
- Extract tools/tauri/process-cleanup.ts (186 lines) from common.ts
- Slim common.ts from ~650 to 286 lines (-56%)
- Unify parseRuntimeMode in grid.ts, used by build.ts and measure-tauri-package.ts

P2 - Cleanup:
- Mark legacy materialize-backend-runtime.ts and smoke-backend-runtime.ts as @deprecated

Verification: typecheck passes, 132 tests pass
