+++ tools/shared/net.ts
@@ -1,0 +1,26 @@
+import { createServer } from 'node:net';
+
+export async function reserveOpenPort(): Promise<number> {
+  return await new Promise<number>((resolvePromise, rejectPromise) => {
+    const server = createServer();
+
+    server.on('error', rejectPromise);
+    server.listen(0, '127.0.0.1', () => {
+      const address = server.address();
+      if (!address || typeof address === 'string') {
+        rejectPromise(new Error('Failed to reserve an open port.'));
+        return;
+      }
+
+      const { port } = address;
+      server.close((error) => {
+        if (error) {
+          rejectPromise(error);
+          return;
+        }
+
+        resolvePromise(port);
+      });
+    });
+  });
+}
